Rewrite authorization to use OAuth claims

+++ src/main/kotlin/dk/cachet/carp/webservices/data/controller/DataStreamController.kt
@@ -4,8 +4,9 @@ import dk.cachet.carp.data.application.DataStreamService
 import dk.cachet.carp.data.infrastructure.DataStreamServiceRequest
 import dk.cachet.carp.webservices.common.configuration.internationalisation.service.MessageBase
 import dk.cachet.carp.webservices.common.exception.responses.BadRequestException
+import dk.cachet.carp.webservices.security.authorization.Claim
+import dk.cachet.carp.webservices.security.authorization.service.AuthorizationService
 import io.swagger.v3.oas.annotations.Operation
-import kotlinx.coroutines.runBlocking
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import org.springframework.http.HttpStatus
@@ -17,9 +18,9 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 class DataStreamController(
-        private val validationMessages: MessageBase,
-       // private val dataStreamPermissionChecker: DataStreamPermissionChecker,
-        private val dataStreamService: DataStreamService
+    private val validationMessages: MessageBase,
+    private val dataStreamService: DataStreamService,
+    private val authorizationService: AuthorizationService
 )
 {
     companion object
@@ -34,32 +35,42 @@ class DataStreamController(
 
     @PostMapping(value = [DATA_STREAM_SERVICE])
     @Operation(tags = ["dataStream/getDataStream.json"])
-    fun getDataStream(
+    suspend fun getDataStream(
             @RequestBody request: DataStreamServiceRequest<*>
-    ): ResponseEntity<Any> = runBlocking {
-        return@runBlocking when (request)
+    ): ResponseEntity<Any> =
+        when (request)
         {
             is DataStreamServiceRequest.OpenDataStreams -> {
-               LOGGER.info("Start POST: $DATA_STREAM_SERVICE -> OpenDataStreams")
+                authorizationService.require( Claim.ManageDeployment( request.configuration.studyDeploymentId ) )
+
+                LOGGER.info("Start POST: $DATA_STREAM_SERVICE -> OpenDataStreams")
                 val result = dataStreamService.openDataStreams(request.configuration)
                 ResponseEntity.status(HttpStatus.CREATED).body(result)
             }
             is DataStreamServiceRequest.AppendToDataStreams -> {
+                authorizationService.require( Claim.InDeployment( request.studyDeploymentId ) )
+
                 LOGGER.info("Start POST: $DATA_STREAM_SERVICE -> AppendToDataStreams")
                 val result = dataStreamService.appendToDataStreams(request.studyDeploymentId, request.batch)
                 ResponseEntity.ok(result)
             }
             is DataStreamServiceRequest.GetDataStream -> {
+                authorizationService.require( Claim.InDeployment( request.dataStream.studyDeploymentId ) )
+
                 LOGGER.info("Start POST: $DATA_STREAM_SERVICE -> GetDataStream")
                 val result = dataStreamService.getDataStream(request.dataStream, request.fromSequenceId, request.toSequenceIdInclusive)
                 ResponseEntity.ok(result)
             }
             is DataStreamServiceRequest.CloseDataStreams -> {
+                authorizationService.require( request.studyDeploymentIds.map { Claim.ManageDeployment( it ) }.toSet() )
+
                 LOGGER.info("Start POST: $DATA_STREAM_SERVICE -> CloseDataStreams")
                 val result = dataStreamService.closeDataStreams(request.studyDeploymentIds)
                 ResponseEntity.ok(result)
             }
             is DataStreamServiceRequest.RemoveDataStreams -> {
+                authorizationService.require( request.studyDeploymentIds.map { Claim.ManageDeployment( it ) }.toSet() )
+
                 LOGGER.info("Start POST: $DATA_STREAM_SERVICE -> RemoveDataStreams")
                 val result = dataStreamService.removeDataStreams(request.studyDeploymentIds)
                 ResponseEntity.ok(result)
@@ -69,5 +80,4 @@ class DataStreamController(
                 throw BadRequestException(validationMessages.get("data.stream.service.request.error", request))
             }
         }
-    }
 }
